MA-10434: add secure playback feature for 2 input buffers.
Reviewed-on: http://androidsource.ap.freescale.net/project/3303

+++ media/libmediaplayerservice/nuplayer/NuPlayerRenderer.cpp
@@ -1410,7 +1410,7 @@ void NuPlayer::Renderer::onDrainVideoQueue() {
         tooLate = isTooLate(realTimeUs,nowUs);
         if (tooLate) {
             ALOGV("video late by %" PRId64 " us (%.2f secs)",
-                 (long long)mVideoLateByUs, mVideoLateByUs / 1E6);
+                 mVideoLateByUs, mVideoLateByUs / 1E6);
         } else {
             int64_t mediaUs = 0;
             mMediaClock->getMediaTime(realTimeUs, &mediaUs);
